% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.aMTM}
\alias{plot.aMTM}
\title{Plot the chain of an aMTM object}
\usage{
\method{plot}{aMTM}(aMTMobject, vars, type, color, pairs, prop.density,
  ...)
}
\arguments{
\item{aMTMobject}{An object of class \code{aMTM}.}

\item{vars}{A vector of indices of the variables to plot. A maximum of 10 variables is accepted and the default 
is all variables of \code{X} up to the first 10.}

\item{type}{Whether to plot points (\code{'p'}), line segments between points (\code{'l'}) or both (\code{'l'}) 
where the upper triangular matrix contains the trace plots and the lower diagonal matrix contains the scatter plot. 
Only applies when \code{pairs==TRUE}; default is \code{'p'}.}

\item{color}{Whether to plot the points/lines colored with the corresponding proposal density. 
In the case of points, the color represent the density used to get to the point; in the case of lines, 
the color represent the density used for the jump. Default is \code{TRUE}.}

\item{pairs}{Whether to plot the variables by pairs or not. Default is \code{TRUE}.}

\item{prop.density}{Whether to add the resulting densities to the plots or not. Default is \code{TRUE}.}

\item{...}{Additionnal graphical parameters.}
}
\description{
An object of class aMTM contains the resulting chain \code{X} as well as the selected proposal at each iteration and the resulting
proposal densities. This function produces various plots to inspect the output of an aMTM object.
}
\details{
When \code{pairs=TRUE}, the 2-D marginal plots of \code{X} for each pair of variables in \code{vars} are plotted. On the diagonal, the
1-D marginal densities are plotted. When \code{pairs=FALSE}, the trace plot and the 1-D marginal plot are produced for each varaibles in 
\code{vars}.
}
\examples{

library(aMTM)
# Banana log-density with parameter B and a
p <- function(x, p) apply(x,1,function(x) -x[1]^2/(2*p$a^2) - 1/2*(x[2]+p$B*x[1]^2-p$B*p$a^2)^2)
# setup
set.seed(1)
N<-1e5;K<-3
B<-0.04;a<-8
# aMTM sampling with ASWAM update
mcmc <- aMTM(target=p, N=N, K=K, x0=c(0,0), parms=list(a=a,B=B), burnin=0.1)

#plot the pairs showing color
plot.aMTM(mcmc, color=T)
#plot the pairs showing jumps and color
plot.aMTM(mcmc, type='l', color=T)
#plot the marginals with colors
plot.aMTM(mcmc, pairs=F, color=T)

}
\author{
Simon Fontaine, \email{fontaines@dms.umontreal.ca}
}
