% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stats.aMTM}
\alias{stats.aMTM}
\title{Statistics of an aMTM object}
\usage{
stats.aMTM(aMTMobject)
}
\arguments{
\item{aMTMobject}{An object of class \code{aMTM}.}
}
\value{
A vector containing the following statistics:

\item{\code{time}}{The computing time.}
\item{\code{msjd}}{The Mean Sqaured Jumping Distance.}
\item{\code{msejd}}{The Mean Sqaured Euclidian Jumping Distance (using the sample variance in the Mahalanobis distance).}
\item{\code{act}}{The Frobenius norm of the multivariate ACT of the chain.}
\item{\code{ess}}{The multivariate ESS of the chain as described by Vatset al. (2015).}
\item{\code{acc.rate}}{The acceptance rate of the chain.}
}
\description{
Summary statistics for the output chain of an aMTM object.
}
\examples{

library(aMTM)
# Banana log-density with parameter B and a
p <- function(x, p) apply(x,1,function(x) -x[1]^2/(2*p$a^2) - 1/2*(x[2]+p$B*x[1]^2-p$B*p$a^2)^2)
# setup
set.seed(1)
N<-1e5;K<-3
B<-0.04;a<-8
# aMTM sampling with ASWAM update
mcmc <- aMTM(target=p, N=N, K=K, x0=c(0,0), parms=list(a=a,B=B), burnin=0.1)

stats.aMTM

}
\references{
Fontaine, S. and Bedard, M. (2019). "An Adaptive Multiple-Try Metropolis algorithm". To be submitted.

Vats, D., Flegal, J. M., and, Jones, G. L. (2015). "Multivariate Output Analysis for Markov chain Monte Carlo". arXiv preprint arXiv:1512.07713.
}
\author{
Simon Fontaine, \email{fontaines@dms.umontreal.ca}
}
